Add afterCreateOrUpdate method to API similar
This is similar to the afterDelete, and very useful when wanting to
generate a response where the resource has already been commited to the
storage.

This particularly solves a problem I experienced while using SSE and
uploading attached files using [cody-code-wy/babyapiFileUploadParser](https://github.com/cody-code-wy/babyapiFileUploadParser).
The event would cause the page to get updated too early and the attempt
ot embed the image would fail with a 404. This kind of situation can be
avoided completely using this extra hook by only sending the events
after everything has been saved and added to the api storage.